* chore(version): set 0.14.3.dev1 version (#1112) (#1113)

* Depend on development version of lambda-builders for dev builds (#1111)

* Depend on development version of lambda-builders for dev builds

* Adding prod.txt to manifest

* Splitting dev and tool dependencies

* fix(build): Resolve path after .aws-sam is created (#1110)

* fix(build): Resolve path after .aws-sam is created

* fix: build (make pr)

* Design and implementation for producing debug build artifacts (#1095)

* design: Initial Design for producing debug artifacts

* initial implementation

* Adding unit tests

* Integration test with debug build mode

* Adjust Design doc and add keyword arg to a call

* fix(dotnet): init template fixes (#1117)

* chore(version): set 0.15.0 (#1125)

* Revert "Depend on development version of lambda-builders for dev builds (#1111)" (#1128)

This reverts commit 7e9de790e23791ba176faff2030286db4007e503.

* Bumping to Lambda Builders 0.3.0 (#1129)

Bumping to Lambda Builders 0.3.0

* fix(func-tests): add dependency manager param (#1130)

+++ samcli/commands/build/build_context.py
@@ -26,6 +26,7 @@ class BuildContext(object):
                  template_file,
                  base_dir,
                  build_dir,
+                 mode,
                  manifest_path=None,
                  clean=False,
                  use_container=False,
@@ -42,6 +43,7 @@ class BuildContext(object):
         self._parameter_overrides = parameter_overrides
         self._docker_network = docker_network
         self._skip_pull_image = skip_pull_image
+        self._mode = mode
 
         self._function_provider = None
         self._template_dict = None
@@ -73,21 +75,16 @@ class BuildContext(object):
 
     @staticmethod
     def _setup_build_dir(build_dir, clean):
+        build_path = pathlib.Path(build_dir)
 
-        # Get absolute path
-        build_dir = str(pathlib.Path(build_dir).resolve())
-
-        if not pathlib.Path(build_dir).exists():
-            # Build directory does not exist. Create the directory and all intermediate paths
-            os.makedirs(build_dir, BuildContext._BUILD_DIR_PERMISSIONS)
-
-        if os.listdir(build_dir) and clean:
-            # Build folder contains something inside. Clear everything.
+        if build_path.exists() and os.listdir(build_dir) and clean:
+            # build folder contains something inside. Clear everything.
             shutil.rmtree(build_dir)
-            # this would have cleared the parent folder as well. So recreate it.
-            os.mkdir(build_dir, BuildContext._BUILD_DIR_PERMISSIONS)
 
-        return build_dir
+        build_path.mkdir(mode=BuildContext._BUILD_DIR_PERMISSIONS, parents=True, exist_ok=True)
+
+        # ensure path resolving is done after creation: https://bugs.python.org/issue32434
+        return str(build_path.resolve())
 
     @property
     def container_manager(self):
@@ -127,3 +124,7 @@ class BuildContext(object):
             return os.path.abspath(self._manifest_path)
 
         return None
+
+    @property
+    def mode(self):
+        return self._mode
